chore: 게시물, 댓글을 행위 중심으로 커뮤니티라는 하나의 BC로 통합 #84
