+++ src/main/java/com/carumuch/capstone/community/infrastructure/BoardImageRepository.java
@@ -1,6 +1,6 @@
-package com.carumuch.capstone.board.repository;
+package com.carumuch.capstone.community.infrastructure;
 
-import com.carumuch.capstone.board.model.BoardImage;
+import com.carumuch.capstone.community.domain.BoardImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BoardImageRepository extends JpaRepository<BoardImage, Long> {
