+++ src/main/java/com/carumuch/capstone/community/infrastructure/BoardRepository.java
@@ -1,6 +1,6 @@
-package com.carumuch.capstone.board.repository;
+package com.carumuch.capstone.community.infrastructure;
 
-import com.carumuch.capstone.board.model.Board;
+import com.carumuch.capstone.community.domain.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BoardRepository extends JpaRepository<Board, Long> {
