+++ src/main/java/com/carumuch/capstone/community/infrastructure/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.carumuch.capstone.comment.repository;
+package com.carumuch.capstone.community.infrastructure;
 
-import com.carumuch.capstone.comment.model.Comment;
+import com.carumuch.capstone.community.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
