+++ src/main/java/com/carumuch/capstone/community/presentation/dto/BoardReqDto.java
@@ -1,4 +1,4 @@
-package com.carumuch.capstone.board.dto;
+package com.carumuch.capstone.community.presentation.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
@@ -6,27 +6,25 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.util.List;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@Schema(name = "BoardModifyReqDto: 게시글 수정 요청 Dto ")
-public class BoardModifyReqDto {
+@Schema(name = "BoardReqDto: 게시글 작성 요청 Dto")
+public class BoardReqDto {
+    @Schema(description = "게시글 제목 입력", example = "boardTitle")
     private String boardTitle;
+    @Schema(description = "게시글 내용 입력", example = "boardContent")
     private String boardContent;
     private List<MultipartFile> boardImage;
-    private List<Long> deleteImage;
 
 
     @Builder
-    public BoardModifyReqDto(String boardTitle,String boardContent,List<MultipartFile> boardImage,List<Long> deleteImage){
+    public BoardReqDto(String boardTitle,String boardContent,List<MultipartFile> boardImage){
         this.boardTitle = boardTitle;
         this.boardContent = boardContent;
         this.boardImage = boardImage;
-        this.deleteImage = deleteImage;
-
     }
 
 }
